IS-7 : Player Death / Sanctuary Point / Particle & Sound

+++ Assets/Scripts/Monsters/Monster.cs
@@ -25,7 +25,7 @@ public class Monster : MonoBehaviour
     public int meleeDamage;
     public int damageMin;
     public int damageMax;
-    public int experience;
+    public float experience;
 
     // Monster Skills
     public int attack;
